+++ app/lib/telegram_api/requests/messages.rb
@@ -3,12 +3,17 @@
 module TelegramApi
   module Requests
     module Messages
-      def send_message(bot_secret:, chat_id:, text:)
+      def send_message(bot_secret:, chat_id:, text:, reply_to_message_id: nil, parse_mode: 'HTML')
         post(
-          path: "bot#{bot_secret}/sendMessage?chat_id=#{chat_id}",
+          path: "bot#{bot_secret}/sendMessage",
           body: {
             text: text
           },
+          params: {
+            chat_id: chat_id,
+            reply_to_message_id: reply_to_message_id,
+            parse_mode: parse_mode
+          }.compact,
           headers: headers
         )
       end