+++ spec/commands/webhooks_context/telegram/receive_chat_member_webhook_command_spec.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-describe WebhooksContext::ReceiveTelegramChatMemberWebhookCommand do
+describe WebhooksContext::Telegram::ReceiveChatMemberWebhookCommand do
   subject(:command_call) { described_class.new.call({ chat_member: chat_member }) }
 
-  let(:handler) { Charkeeper::Container.resolve('services.webhooks_context.handle_telegram_chat_member_webhook') }
+  let(:handler) { Charkeeper::Container.resolve('services.webhooks_context.telegram.handle_chat_member_webhook') }
 
   before { allow(handler).to receive(:call) }
 
@@ -13,6 +13,22 @@ describe WebhooksContext::ReceiveTelegramChatMemberWebhookCommand do
         from: { first_name: 'First', last_name: 'Last', username: 'User', language_code: 'en' },
         chat: { id: '' },
         new_chat_member: { status: 'kicked' }
+      }
+    end
+
+    it 'does not call handler' do
+      command_call
+
+      expect(handler).not_to have_received(:call)
+    end
+  end
+
+  context 'with group message' do
+    let(:chat_member) do
+      {
+        from: { first_name: 'First', last_name: 'Last', username: 'User', language_code: 'en' },
+        chat: { id: -1 },
+        new_chat_member: { status: 'member' }
       }
     end