+++ spec/services/webhooks_context/telegram/handle_chat_member_webhook_service_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-describe WebhooksContext::HandleTelegramChatMemberWebhookService do
+describe WebhooksContext::Telegram::HandleChatMemberWebhookService do
   subject(:service_call) { described_class.new.call(chat_member: chat_member) }
 
   let(:chat_member) do